refactor: extract ADT text-search contract
