+++ src/adt/xml-parser.ts
@@ -39,7 +39,6 @@ import type {
   RevisionInfo,
   RevisionListResult,
   ServerDrivenObjectMetadata,
-  SourceSearchResult,
   TransactionInfo,
 } from './types.js';
 
@@ -603,159 +602,6 @@ function normalizeDiscoveryPath(href: string): string | undefined {
   }
 
   return path;
-}
-
-/** Pull `objectName` out of an ADT proxy-URI mapping.
- *
- *  Text-search results do not carry the object name as an attribute; it is
- *  URL-encoded inside the `content` query parameter, e.g.
- *  `…?id=sris.objectType&content=objectType%3aCLASI%2cobjectName%3aZCL_FOO%3d%3d%3dCCAU`.
- *  Class includes are padded to 30 characters with `=` and given a 4-character
- *  suffix (CCAU, CCIMP, …); the padding is dropped so the caller sees the
- *  enclosing class it recognises rather than `ZCL_FOO=========CCAU`. */
-function parseProxyUriObjectName(uri: string): string {
-  if (!uri) return '';
-  const content = /[?&]content=([^&]*)/.exec(uri.replace(/&amp;/g, '&'));
-  if (!content) return '';
-  let decoded = content[1];
-  try {
-    decoded = decodeURIComponent(decoded);
-  } catch {
-    // Malformed percent-encoding — fall through with the raw value.
-  }
-  const named = /objectName:([^,#]+)/i.exec(decoded);
-  const raw = named ? named[1] : decoded.split('#')[0];
-  return raw.replace(/=+.*$/, '').trim();
-}
-
-/** Pull the 1-based line number out of a textLine proxy URI.
- *
- *  Two encodings occur, depending on the editor the hit maps to: source-backed
- *  objects carry an `#start=51,0;end=51,0` fragment, while others (XSLT, for
- *  one) carry a `,position:25` segment instead. Returns 0 when neither is
- *  present, which keeps the snippet usable even without a line number. */
-function parseTextLineNumber(uri: string): number {
-  if (!uri) return 0;
-  let decoded = uri.replace(/&amp;/g, '&');
-  try {
-    decoded = decodeURIComponent(decoded);
-  } catch {
-    // Malformed percent-encoding — fall through with the raw value.
-  }
-  const match = /#start=(\d+)/.exec(decoded) ?? /\bposition:(\d+)/.exec(decoded);
-  return match ? Number(match[1]) : 0;
-}
-
-/** Strip the `<b>` markers ADT wraps around the hit and collapse the surrounding
- *  whitespace, which arrives with the server's own line breaks and ellipses. */
-function cleanTextSearchSnippet(raw: unknown): string {
-  return decodeXmlEntities(String(raw ?? ''))
-    .replace(/<\/?b>/gi, '')
-    .replace(/\s+/g, ' ')
-    .trim();
-}
-
-/**
- * Parse ADT source/text search results.
- *
- * The live `informationsystem/textsearch` endpoint answers with a
- * `textsearch:textSearchResult` document:
- *
- *   <textsearch:textSearchObject uri="…objectName%3aZCL_FOO…" parentUri="…" isResult="false">
- *     <textsearch:adtMainObject adtcore:name="Test Classes" adtcore:type="CLAS/I"/>
- *     <textsearch:textLines>
- *       <textsearch:textLine uri="…CCAU%23start%3d51%2c0%3bend%3d51%2c0">
- *         <textsearch:content>… &lt;b&gt;hit&lt;/b&gt; = lv_ok ).…</textsearch:content>
- *
- * Neither the object name nor the line number is an attribute — both are
- * URL-encoded inside proxy-URI mappings and have to be decoded out. The
- * objectReference and Atom shapes are kept as fallbacks for other releases.
- */
-export function parseSourceSearchResults(xml: string): SourceSearchResult[] {
-  const parsed = parseXml(xml);
-  const results: SourceSearchResult[] = [];
-
-  // Live textsearch format
-  const searchObjects = findDeepNodes(parsed, 'textSearchObject');
-  if (searchObjects.length > 0) {
-    for (const obj of searchObjects) {
-      const matches = findDeepNodes(obj, 'textLine').map((line) => ({
-        line: parseTextLineNumber(String(line['@_uri'] ?? '')),
-        snippet: cleanTextSearchSnippet(line.content),
-      }));
-
-      // The response is a tree: alongside the objects that actually contain the
-      // text it carries their ancestors (owning class, package) as nodes with no
-      // textLines. Those are navigation scaffolding rather than search hits, and
-      // they do not count against the server's paging window either.
-      if (matches.length === 0) continue;
-
-      const uri = decodeXmlEntities(String(obj['@_uri'] ?? ''));
-      const mainObject = (obj.adtMainObject ?? {}) as Record<string, unknown>;
-      results.push({
-        objectType: String(mainObject['@_type'] ?? ''),
-        // Deliberately not parentUri: for a class include that is the class, but
-        // for a top-level object it is the package — which would report every
-        // program in $TMP as "$TMP".
-        objectName: parseProxyUriObjectName(uri) || String(mainObject['@_name'] ?? ''),
-        uri,
-        matches,
-      });
-    }
-    return results;
-  }
-
-  // Try objectReferences format (similar to quickSearch)
-  const refs = getNestedArray(parsed, 'objectReferences', 'objectReference');
-  if (refs.length > 0) {
-    for (const ref of refs) {
-      const matchNodes = findDeepNodes(ref, 'textSearchResult');
-      const matches = matchNodes.map((m: Record<string, unknown>) => ({
-        line: Number(m['@_line'] ?? 0),
-        snippet: String(m['@_snippet'] ?? m['#text'] ?? ''),
-      }));
-      results.push({
-        objectType: String(ref['@_type'] ?? ''),
-        objectName: String(ref['@_name'] ?? ''),
-        uri: String(ref['@_uri'] ?? ''),
-        matches,
-      });
-    }
-    return results;
-  }
-
-  // Try Atom feed format
-  const entries = getNestedArray(parsed, 'feed', 'entry');
-  for (const entry of entries) {
-    const uri = String(entry.id ?? entry['@_href'] ?? '');
-    const title = String(entry.title ?? '');
-    results.push({
-      objectType: '',
-      objectName: title || uri.split('/').pop() || '',
-      uri,
-      matches: [],
-    });
-  }
-
-  // Fallback: try to find any matching nodes
-  if (results.length === 0) {
-    const nodes = findDeepNodes(parsed, 'match');
-    for (const node of nodes) {
-      results.push({
-        objectType: String(node['@_type'] ?? ''),
-        objectName: String(node['@_name'] ?? node['@_objectName'] ?? ''),
-        uri: String(node['@_uri'] ?? ''),
-        matches: [
-          {
-            line: Number(node['@_line'] ?? 0),
-            snippet: String(node['@_snippet'] ?? node['#text'] ?? ''),
-          },
-        ],
-      });
-    }
-  }
-
-  return results;
 }
 
 /**
